requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Deprecated==1.2.18
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.7
ordered-set==4.1.0
packaging==24.2
passlib==1.7.4
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
pytz==2025.1
rich==13.9.4
SQLAlchemy==2.0.37
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
pytest==8.2.1
python-dotenv==1.0.1
